interaction/requirements.txt
pwntools
sleekxmpp
